+++ Nova/Actions/ChangeEndsAt.php
@@ -3,13 +3,14 @@
 namespace App\Nova\Actions;
 
 use App\Models\User;
-use Laravel\Nova\Fields\Date;
 use Illuminate\Bus\Queueable;
-use Laravel\Nova\Actions\DestructiveAction;
+use Laravel\Nova\Fields\Date;
 use Illuminate\Support\Collection;
 use Laravel\Nova\Fields\ActionFields;
+use Illuminate\Support\Facades\Artisan;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use Laravel\Nova\Actions\DestructiveAction;
 
 class ChangeEndsAt extends DestructiveAction
 {
@@ -27,7 +28,7 @@ class ChangeEndsAt extends DestructiveAction
         foreach($models as $subscription) {
             $subscription->ends_at = $fields['ends_at'];
 
-            $subscription->save();                        
+            $subscription->save();            
         }
     }